Processor renaming for consistency
